model 업데이트
