+++ src/models/role/role.js
@@ -1,17 +1,29 @@
+import OrganizationModel from "../organization/organization.js";
+import PermissionGroupModel from "../permission/permission_group.js";
+
 export default (sequelize, Sequelize) => {
+	const Organization = OrganizationModel(sequelize, Sequelize);
+	const PermissionGroup = PermissionGroupModel(sequelize, Sequelize);
+
 	return sequelize.define(
-		"Permissions",
+		"Role",
 		{
 			id: {
 				type: Sequelize.INTEGER,
 				primaryKey: true,
 				autoIncrement: true,
-				comment: "권한 고유 식별자",
+				comment: "역할 고유 식별자",
 			},
-			permission_code: {
+			name: {
 				type: Sequelize.STRING(20),
 				allowNull: false,
-				comment: "권한 코드",
+				comment: "역할 이름",
+			},
+			is_deleted: {
+				type: Sequelize.BOOLEAN,
+				allowNull: false,
+				defaultValue: false,
+				comment: "삭제 여부 (Y: 삭제됨, N: 활성 상태)",
 			},
 			created_at: {
 				type: Sequelize.DATE,
@@ -27,11 +39,11 @@ export default (sequelize, Sequelize) => {
 			},
 		},
 		{
-			tableName: "permissions",
+			tableName: "role",
 			timestamps: true,
 			createdAt: "created_at",
 			updatedAt: "updated_at",
-			comment: "권한 테이블",
+			comment: "역할 정보를 관리하는 테이블",
 		}
 	);
 };
