+++ src/models/season/season.js
@@ -12,15 +12,15 @@ export default (sequelize, Sequelize) => {
 				autoIncrement: true,
 				comment: "시즌 고유 식별자",
 			},
-			season_name: {
+			name: {
 				type: Sequelize.STRING(20),
 				allowNull: false,
 				comment: "시즌 이름",
 			},
 			is_deleted: {
-				type: Sequelize.CHAR(1),
+				type: Sequelize.BOOLEAN,
 				allowNull: false,
-				defaultValue: "N",
+				defaultValue: false,
 				comment: "삭제 여부 (Y: 삭제됨, N: 활성 상태)",
 			},
 			created_at: {
@@ -35,38 +35,13 @@ export default (sequelize, Sequelize) => {
 				defaultValue: Sequelize.NOW,
 				comment: "데이터 최종 수정 일시",
 			},
-			creator_id: {
-				type: Sequelize.INTEGER,
-				allowNull: false,
-				comment: "데이터를 생성한 사용자의 ID",
-			},
-			updater_id: {
-				type: Sequelize.INTEGER,
-				allowNull: false,
-				comment: "데이터를 마지막으로 수정한 사용자의 ID",
-			},
-			creator_ip: {
-				type: Sequelize.STRING(45),
-				allowNull: false,
-				comment: "데이터를 생성한 사용자의 IP 주소",
-			},
-			updater_ip: {
-				type: Sequelize.STRING(45),
-				allowNull: false,
-				comment: "데이터를 마지막으로 수정한 사용자의 IP 주소",
-			},
-			access_service_id: {
-				type: Sequelize.STRING(50),
-				allowNull: true,
-				comment: "데이터 유입 채널",
-			},
 		},
 		{
 			tableName: "season",
 			timestamps: true,
 			createdAt: "created_at",
 			updatedAt: "updated_at",
-			comment: "시즌 정보를 관리하는 테이블",
+			comment: "회기 테이블",
 		}
 	);
 };
